comments, put back some code Brian removed (need to ask him about that)

+++ mojos/src/main/java/gov/va/isaac/mojos/dbBuilder/Shutdown.java
@@ -1,11 +1,11 @@
 package gov.va.isaac.mojos.dbBuilder;
 
 import gov.va.isaac.AppContext;
-
+import java.io.File;
 import java.util.List;
-
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
+import org.ihtsdo.otf.tcc.api.io.FileIO;
 import org.ihtsdo.otf.tcc.api.store.TerminologyStoreDI;
 import org.ihtsdo.otf.tcc.model.index.service.IndexerBI;
 
@@ -64,15 +64,17 @@ public class Shutdown extends AbstractMojo {
         indexer.closeWriter();
       }
 
-/*      if (moveToReadOnly) {
-          Thread.sleep(5000);
+      if (moveToReadOnly) {
+        //TODO Ask Brian why he disabled this - probably on a dev call and ask Keith about it too, since it came from him initially.
+        //TODO OTF Bug - figure out why on earth we need this arbitrary sleep.
+        Thread.sleep(5000);
         getLog().info("moving mutable to read-only");
         File readOnlyDir = new File(bdbFolderLocation, "read-only");
         FileIO.recursiveDelete(readOnlyDir);
         File mutableDir = new File(bdbFolderLocation, "mutable");
         mutableDir.renameTo(readOnlyDir);
       }
-*/      
+      
       getLog().info("Done shutting down terminology store");
     } catch (Exception e) {
       throw new MojoExecutionException("Database build failure", e);
